+++ src/Jieba.hpp
@@ -53,7 +53,10 @@ class Jieba {
   const HMMModel* GetHMMModel() const {
     return &model_;
   }
-
+ 
+  void SetQuerySegmentThreshold(size_t len) {
+    query_seg_.SetMaxWordLen(len);
+  }
  private:
   DictTrie dict_trie_;
   HMMModel model_;